+++ data/src/App.svelte
@@ -1,6 +1,9 @@
 <script lang="ts">
   import { Router } from 'sv-router';
   import 'sv-router/generated';
+  import SetupGate from '$lib/components/SetupGate.svelte';
 </script>
 
-<Router />
+<SetupGate>
+  <Router />
+</SetupGate>